Updated to version 12.6.0-beta002 (code 479).
* In app/build.gradle.kts:

    - Changed "jvmTarget" to use the updated format in kotlin ->
      compilerOptions.

* In gradle.properties:

    - Removed usage of buildconfig=true since it was no longer
      necessary.

* In AndroidManifest.xml:

    - Removed targetApi="n" on RECEIVE_BOOT_COMPLETED since that is now
      the min SDK.

* In (googleplay) support/NacSupportSetting.kt:

    - Updated the javadoc string for onProductDetailsResponse().

* In alarm/activealarm/NacActiveAlarmNotification.kt:

    - Changed @TargetApi to @RequiresApi on createChannel().

* In alarm/activealarm/NacSkipAlarmNotification.kt:

    - Changed @TargetApi to @RequiresApi on createChannel().

* In alarm/activealarm/NacSwipeAnimationHandler.kt:

    - Changed View visibility checks to use the kotlin extension.

    - Changed exception variable to _.

* In alarm/activealarm/NacSwipeLayoutHandler.kt:

    - Changed String to Uri conversion to use kotlin extension.

    - Changed exception variable to _.

* In alarm/options/audiosource/NacAudioSourceDialog.kt:

    - Changed RadioGroup child check to use kotlin extension.

* In alarm/options/mediapicker/NacMediaPickerFragment.kt:

    - Changed "mediaArtist", "mediaTitle", and "localMediaPath" to
      public so that children can modify them.

    - Moved majority of logic in onOkClicked() to children as it was
      more appropriate there. Also, was doing some testing on API 24 and
      it was treating a ringtone as a file, so it was opening the Browse
      tab first when it should not have been. This is because the
      onOkClicked() logic was generalized and it was treating the path
      as a File path.

      All that is left now is the cleanup and finishing the activity
      which should be generalized because each child will do the same
      thing.

* In alarm/options/mediapicker/music/NacMusicPickerFragment.kt:

    - Changed String to Uri conversion to use kotlin extension.

    - Moved logic where the "media*" variables are set in
      onOkClicked() from NacMediaPickerFragment to here. Then the parent
      method is called.

      TODO: Test that this works properly when a directory is clicked.

* In alarm/options/mediapicker/ringtone/NacRingtonePickerFragment.kt:

    - Changed String to Uri conversion to use kotlin extension.

    - Moved logic where the "media*" variables are set in
      onOkClicked() from NacMediaPickerFragment to here. Then the parent
      method is called.

* In alarm/options/missedalarm/NacMissedAlarmNotification.kt:

    - Changed @TargetApi to @RequiresApi on createChannel().

* In alarm/options/name/NacNameDialog.kt:

    - Optimized imports.

* In alarm/options/snoozeoptions/NacSnoozeOptionsDialog.kt:

    - Optimized imports.

* In alarm/options/upcomingreminder/NacUpcomingReminderNotification.kt:

    - Changed @TargetApi to @RequiresApi on createChannel().

* In alarm/options/vibrate/NacVibrator.kt:

    - Removed @TargetApi because it seemed like it was not necessary.

* In card/NacCardHolder.kt:

    - Changed View visibility checks to use the kotlin extension.

* In db/NacAlarmDatabase.kt:

    - Removed check on version code N as that is now the min SDK.

    - Changed exception variable to _.

* In db/NacOldDatabase.kt:

    - Changed database transaction to use the kotlin extension.

    - Changed String to Uri conversion to use kotlin extension.

* In main/NacMainActivity.kt:

    - Changed String to Uri conversion to use kotlin extension.

    - Added TODO to test intent, to add checks later so it is called
      safely.

    - Changed Menu size check to use kotlin extension.

* In settings/NacAppearanceSettingFragment.kt:

    - Removed check on version code N as that is now the min SDK.

* In settings/NacGeneralSettingFragment.kt:

    - Removed check on version code N as that is now the min SDK.

* In settings/NacMainSettingActivity.kt:

    - Changed ColorDrawable to use kotlin extension.

* In shared/NacSharedPreferences.kt:

    - Changed save*() to use shared preference edit kotlin extension.

* In system/file/NacFile.kt:

    - Updated code to use Environment.getExternalStorageDirectory() to
      get the /storage/emulated/... path

    - Changed exception variable to _.

* In system/file/NacFileTree.kt:

    - Removed @TargetApi because it seemed like it was not necessary.

    - Changed exception variable to _.

* In system/file/browser/NacFileBrowserRepository.kt:

    - Changed exception variable to _.

* In system/mediaplayer/NacMediaPlayer.kt:

    - Changed String to Uri conversion to use kotlin extension.

* In system/permission/NacPermissionRequestManager.kt:

    - Removed @RequiresApi because it seemed like it was not necessary.

      TODO: Investigate this to make sure it is ok.

* In util/NacContext.kt:

    - Removed check on version code N as that is now the min SDK.

* In util/NacIntent.kt:

    - Changed String to Uri conversion to use kotlin extension.

* In util/media/NacAudioAttributes.kt:

    - Removed unnecessary conversion toFloat().

* In util/media/NacAudioManager.kt:

    - Fixed issue where if the stream was not set, which meant that it
      was set to USE_DEFAULT_STREAM_TYPE, then media would not be able
      to be played. This would happen in API 24 when trying to play a
      ringtone. Now, if this is the case, then the stream is set to
      STREAM_MUSIC by default.

* In util/media/NacMedia.kt:

    - Changed String to Uri conversion to use kotlin extension.

    - Changed exception variable to _.

    - Removed @TargetApi because it seemed like it was not necessary.

* In view/NacView.kt:

    - Removed check on version code N as that is now the min SDK.

* In view/colorpicker/NacColorPickerDialog.kt:

    - Changed String to Color conversion to use kotlin extension.

    - Changed exception variable to _.

* In view/notification/NacNotification.kt:

    - Removed @TargetApi on builder().

    - Changed @TargetApi to @RequiresApi on createChannel().

* In whatsnew/NacWhatsNewDialog.kt:

    - Optimized imports.

* In whatsnew/NacWhatsNewListAdapter.kt:

    - Updated to use String.toSpannedString() which is a custom function
      extension.

* In widget/NacClockWidgetConfigureActivity.kt:

    - Changed View visibility checks to use the kotlin extension.

* In layout/act_alarm_new.xml:

    - Added ignore of "Overdraw".

* In values*/strings.xml:

    - Removed unused resources.

* Removed the following resources:

    - layout/dlg_alarm_days.xml
    - layout/nac_preference_volume.xml
    - mipmap-hdpi/app.webp

+++ app/src/main/kotlin/com/nfcalarmclock/util/media/NacAudioAttributes.kt
@@ -186,7 +186,7 @@ class NacAudioAttributes(
 		volumeLevel = alarm.volume
 
 		// Set the text-to-speech rate and voice
-		speechRate = alarm.ttsSpeechRate.toFloat()
+		speechRate = alarm.ttsSpeechRate
 		voice = alarm.ttsVoice
 
 		return this
